pipelines/original/requirements.txt
opencv-python
pyyaml
progressbar
tqdm
imutils
uuid
numpy
pandas
Pillow
scipy
rospkg
netifaces
mpi4py
